Modified code for u16

+++ tests/tests.rs
@@ -19,7 +19,7 @@ fn test_ranges() {
         (7,11),
         (3,7),
     ];
-    let funcs: Vec<(&str, fn(u64,u32)->u64)> = vec![
+    let funcs: Vec<(&str, fn(u16,u32)->u16)> = vec![
         ("pow_std", pow_std),
         ("pow_std_2opt", pow_std_2opt),
         ("pow_alt", pow_alt),
@@ -31,7 +31,7 @@ fn test_ranges() {
     ];
     for (base, max_exp) in args {
         for exp in 0..=max_exp {
-            let answer = u64::pow(base, exp);
+            let answer = u16::pow(base, exp);
             for (name, func) in &funcs {
                 println!("{}({},{})", name, base, exp);
                 assert_eq!(answer, (*func)(base, exp));
